some final changes

+++ src/Components/Home/HomePage.js
@@ -48,6 +48,8 @@ const HomePage = () => {
   };
 
   const handleDelete = async (id) => {
+    const confirmDelete = window.confirm("Are you sure you want to delete this record?");
+    if(!confirmDelete) return;
     try {
       await axios.delete(`${baseUrl}/${id}`);
       notify();
